refactor(bds): remove deprecated ServiceBadge — ServiceTag is canonical (closes #572)
ServiceBadge has been `@deprecated` with a JSDoc pointer at ServiceTag since
the component-bloat audit (ADR-004). The portal — the only meaningful
consumer — already uses canonical ServiceTag via a local adapter wrapper
(@/components/service-badge), so the BDS deletion has zero consumer blast
radius.

### Architecture: extracted shared symbols

ServiceTag was importing `categoryConfig`, `getServiceIconPath`, types from
the deprecated ServiceBadge.tsx file. Moved those to a new shared module:

  components/ui/ServiceBadge/service-config.ts

Contains: ServiceCategory, ServiceBadgeSize, categoryConfig,
serviceIconOverrides, normalizeServiceName, getServiceIconPath. Both the
package-level exports and ServiceTag continue to consume these from the new
location — no API change for consumers.

### Deleted
- `components/ui/ServiceBadge/ServiceBadge.tsx` (component + interface)
- `components/ui/ServiceBadge/ServiceBadge.css`
- `docs-site/content/docs/components/service-badge.mdx` (deprecation page)

### Updated
- `components/ui/ServiceBadge/index.ts` — drop ServiceBadge / ServiceBadgeProps / ServiceBadgeMode exports
- `components/ui/ServiceBadge/ServiceTag.{tsx,stories.tsx,mdx}` — import from ./service-config; past-tense Migration; add @summary
- `components/ui/PageHeader/PageHeader.tsx` — JSDoc example updated to ServiceTag
- `docs-site/content/docs/components/{meta.json,index.mdx,page-header.mdx,icons.mdx}` — drop nav entry, update references, replace examples
- `scripts/inspector-overrides.json` — drop ServiceBadge override block
- `scripts/build-inspector-manifest.mjs` — skip dirs without `${PascalName}.tsx`
- `scripts/lint-jsdoc.js` — add ServiceBadge to MULTI_EXPORT (canonical = ServiceTag.tsx)
- `manifest/component-axes.json` — regenerated

### bds-manifest improvements (incidental, related)

Adding the `${PascalName}.tsx` guard to the builder also drops vestigial
entries for `Calendar/` (placeholder, not-ready) and `Icons/` (barrel
module, not a component). Confirmed not consumed via any `bds-calendar`
or `bds-icons` class in portal / renew / brikdesigns. Net manifest count:
97 → 96 (#725 AlertBanner) → 93 (this PR + guard).

URL resolution vs live Chromatic /index.json: 92/93 (99%). Only remaining
miss is AnimatedIcon, tracked in #726.

### Out of scope (follow-up)

The `components/ui/ServiceBadge/` directory name still mismatches the
canonical component (ServiceTag). Renaming touches every internal
`'./ServiceBadge'` / `'../ServiceBadge'` import path. Tracked separately.
