--- components/ui/ServiceBadge/ServiceBadge.css
@@ -1,36 +1,0 @@
-@layer bds-components {
-/* ─── ServiceBadge (deprecated — use ServiceTag variant="icon") ── */
-
-.bds-service-badge {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  overflow: hidden;
-}
-
-/* ─── Sizes ──────────────────────────────────────────────────── */
-/* bds-lint-ignore — Figma-driven badge dimensions */
-
-.bds-service-badge--sm { width: 20px; height: 20px; border-radius: var(--border-radius-sm); } /* bds-lint-ignore */
-.bds-service-badge--md { width: 28px; height: 28px; border-radius: var(--border-radius-md); } /* bds-lint-ignore */
-.bds-service-badge--lg { width: 40px; height: 40px; border-radius: var(--border-radius-md); } /* bds-lint-ignore */
-
-/* ─── Category colors ────────────────────────────────────────── */
-
-.bds-service-badge--brand       { background-color: var(--background-service-brand);       color: var(--text-service-brand); }
-.bds-service-badge--marketing   { background-color: var(--background-service-marketing);   color: var(--text-service-marketing); }
-.bds-service-badge--information { background-color: var(--background-service-information); color: var(--text-service-information); }
-.bds-service-badge--product     { background-color: var(--background-service-product);     color: var(--text-service-product); }
-/* `--service` class modifier refers to the back-office service line; tokens were
-   renamed `service-service` → `service-back-office` in #563. Class kept stable
-   for component prop-API compat; follow-up may rename the modifier separately. */
-.bds-service-badge--service     { background-color: var(--background-service-back-office); color: var(--text-service-back-office); }
-
-/* ─── Icon ───────────────────────────────────────────────────── */
-
-.bds-service-badge__icon {
-  object-fit: contain;
-  display: block;
-}
-}
