+++ components/ui/ServiceBadge/ServiceTag.mdx
@@ -85,7 +85,7 @@ import { ServiceTag } from '@brik/bds';
 
 ## Migration from ServiceBadge
 
-`ServiceBadge` is deprecated. Migrate to `ServiceTag variant="icon"`:
+`ServiceBadge` was removed in [#572](https://github.com/brikdesigns/brik-bds/issues/572). Existing usages migrate to `ServiceTag variant="icon"`:
 
 ```tsx
 // Old
