+++ docs-site/content/docs/components/icons.mdx
@@ -23,7 +23,7 @@ BDS registers the Phosphor collection at startup in `.storybook/preview.tsx` via
 - Status / state markers paired with text
 - Anywhere a small (≤ 32px) glyph carries meaning
 
-For larger illustrations and brand marks, use a static SVG asset directly. For non-icon visual primitives ([ServiceBadge](/docs/components/service-badge), avatar fallbacks), use the dedicated component.
+For larger illustrations and brand marks, use a static SVG asset directly. For non-icon visual primitives ([ServiceTag](https://storybook.brikdesigns.com/?path=/docs/components-service-tag--overview), avatar fallbacks), use the dedicated component.
 
 ## Import
 
@@ -140,6 +140,6 @@ export const HeartFill = 'ph:heart-fill';
 
 - [IconButton](/docs/components/button#iconbutton) — accessible wrapper for icon-only buttons
 - [Avatar](/docs/components/avatar) — for person identity
-- [ServiceBadge](/docs/components/service-badge) — for Brik service-line icons
+- [ServiceTag](https://storybook.brikdesigns.com/?path=/docs/components-service-tag--overview) — for Brik service-line icons (use `variant="icon"`)
 - **[Storybook icon catalog](https://storybook.brikdesigns.com/?path=/docs/components-icons--overview)**
 - **External**: [phosphoricons.com](https://phosphoricons.com/) — find icon names
