+++ docs-site/content/docs/components/index.mdx
@@ -11,7 +11,7 @@ BDS components are organized by **job**, not by visual type. The same taxonomy b
 |---|---|
 | **Action** | Button, IconButton, LinkButton, ButtonGroup, FilterBar, FilterButton, FilterToggle, TextLink |
 | **Form** | TextInput, TextArea, PasswordInput, AddressInput, Select, MultiSelect, DatePicker, TimePicker, Checkbox, Radio, Field, FieldGrid, Form |
-| **Indicator** | Badge, Chip, Tag, Counter, Dot, Spinner, Skeleton, BadgeGroup, TagGroup, ServiceBadge |
+| **Indicator** | Badge, Chip, Tag, Counter, Dot, Spinner, Skeleton, BadgeGroup, TagGroup, ServiceTag |
 | **Feedback** | Banner, Toast, Tooltip, EmptyState, ProgressBar |
 | **Control** | Switch, Slider, Stepper, ProgressStepper, Pagination, SegmentedControl, FileUploader |
 | **Addable** | AddableTextList, AddableComboList, AddableEntryList, AddableFieldRowList |
@@ -41,7 +41,7 @@ The indicator family — read-only state communication. For interactive pills, u
   <Card title="Chip" href="/docs/components/chip" description="Interactive pill for filtering, selection, removable tokens. Always wired up." />
   <Card title="Counter" href="/docs/components/counter" description="Numeric count with status colors. Notification badges, item quantities." />
   <Card title="Dot" href="/docs/components/dot" description="Minimal status circles. Six statuses, three sizes, optional pulse." />
-  <Card title="Service badge" href="/docs/components/service-badge" description="Icon-only badge for Brik service categories." />
+  <Card title="Service tag" href="https://storybook.brikdesigns.com/?path=/docs/components-service-tag--overview" description="Icon, text, or icon+text indicator for Brik service categories." />
   <Card title="Skeleton" href="/docs/components/skeleton" description="Loading-state placeholder shapes. Text, circular, rectangular." />
   <Card title="Spinner" href="/docs/components/spinner" description="Circular loading indicator. Inline and container sizes." />
   <Card title="Tag" href="/docs/components/tag" description="Neutral categorization labels. Read-only, four sizes, two appearances." />
