+++ docs-site/content/docs/components/meta.json
@@ -15,7 +15,6 @@
     "chip",
     "counter",
     "dot",
-    "service-badge",
     "skeleton",
     "spinner",
     "tag",
