+++ scripts/inspector-overrides.json
@@ -240,9 +240,6 @@
         "fixed": ["keyboard-nav", "focus-trap-menu", "aria-listbox"]
       }
     },
-    "ServiceBadge": {
-      "introduced_in": "0.1.3"
-    },
     "Sheet": {
       "introduced_in": "0.1.3"
     },
